# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для записи к специалистам",
        "title": "LAPS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "94.247.129.222:8080",
    "basePath": "/api/v1",
    "paths": {
        "/appointments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список записей на консультации с фильтрацией и пагинацией",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Записи"
                ],
                "summary": "Получить список записей",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Лимит записей на странице (по умолчанию 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение (по умолчанию 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID клиента (только для админов)",
                        "name": "client_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID специалиста (только для админов)",
                        "name": "specialist_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Статус записи",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Начальная дата (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конечная дата (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список записей с пагинацией",
                        "schema": {
                            "$ref": "#/definitions/rest.paginatedResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новую запись на консультацию к специалисту",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Записи"
                ],
                "summary": "Создать запись на консультацию",
                "parameters": [
                    {
                        "description": "Данные для записи на консультацию",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateAppointmentDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID созданной записи",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации или выбранное время недоступно",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        },
        "/appointments/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает информацию о записи на консультацию по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Записи"
                ],
                "summary": "Получить запись по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID записи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные записи",
                        "schema": {
                            "$ref": "#/definitions/domain.Appointment"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Запись не найдена",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет информацию о записи на консультацию",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Записи"
                ],
                "summary": "Обновить запись",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID записи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления записи",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateAppointmentDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном обновлении",
                        "schema": {
                            "$ref": "#/definitions/rest.messageResponseType"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации или выбранное время недоступно",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Запись не найдена",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Отменяет запись на консультацию",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Записи"
                ],
                "summary": "Отменить запись",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID записи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешной отмене",
                        "schema": {
                            "$ref": "#/definitions/rest.messageResponseType"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID или ошибка отмены",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Запись не найдена",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Appointment": {
            "type": "object",
            "properties": {
                "appointment_date": {
                    "type": "string"
                },
                "client_id": {
                    "type": "integer"
                },
                "communication_method": {
                    "$ref": "#/definitions/domain.CommunicationMethod"
                },
                "consultation_type": {
                    "$ref": "#/definitions/domain.ConsultationType"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "payment_id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "specialist_id": {
                    "type": "integer"
                },
                "specialization_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/domain.AppointmentStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.AppointmentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "paid",
                "completed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "AppointmentStatusPending",
                "AppointmentStatusPaid",
                "AppointmentStatusCompleted",
                "AppointmentStatusCancelled"
            ]
        },
        "domain.CommunicationMethod": {
            "type": "string",
            "enum": [
                "phone",
                "whatsapp"
            ],
            "x-enum-varnames": [
                "CommunicationMethodPhone",
                "CommunicationMethodWhatsApp"
            ]
        },
        "domain.ConsultationType": {
            "type": "string",
            "enum": [
                "primary",
                "secondary"
            ],
            "x-enum-varnames": [
                "ConsultationTypePrimary",
                "ConsultationTypeSecondary"
            ]
        },
        "domain.CreateAppointmentDTO": {
            "type": "object",
            "required": [
                "appointment_date",
                "communication_method",
                "consultation_type",
                "specialist_id"
            ],
            "properties": {
                "appointment_date": {
                    "type": "string"
                },
                "communication_method": {
                    "enum": [
                        "phone",
                        "whatsapp"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.CommunicationMethod"
                        }
                    ]
                },
                "consultation_type": {
                    "enum": [
                        "primary",
                        "secondary"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.ConsultationType"
                        }
                    ]
                },
                "specialist_id": {
                    "type": "integer"
                },
                "specialization_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.UpdateAppointmentDTO": {
            "type": "object",
            "properties": {
                "appointment_date": {
                    "type": "string"
                },
                "payment_id": {
                    "type": "string"
                },
                "status": {
                    "enum": [
                        "pending",
                        "paid",
                        "completed",
                        "cancelled"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.AppointmentStatus"
                        }
                    ]
                }
            }
        },
...
        "rest.errorResponseBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "rest.messageResponseType": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "rest.paginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}